Moved keyboard input polling outside of the clock time loop. This isn't necessary so long as you're at least getting 1-2 fps (which would result in a typing speed between 60-120 wpm). Even then, you've got more serious things to be thinking about at that point than if all of your input is being tracked, and having it inside the clock time loop just cuts performance for everyone else.
This loop should get removed altogether, but there are a few items left
in it at the moment which would need to get reworked first before it can
be eliminated. In cases where you can't accurately know where something
would be because it's random, we should just scale the random formulas
to take a more drastic adjustment to compensate for that lag.

Signed-off-by: Tametomo <irarice@gmail.com>