Massive revamps of input handling based around judging whether focus has been taken or not. All text input bugs should now be fixed for good, and now we mainly need to work on extending key support for certain dialogs and even widgets.
This commit also gets rid of the old keyboard configuration option for
the OK key. So long as the client knows to behave and wait its turn on
this key, input should be fine.

Probably should place warnings in the keyboard setup tab on using
blocked keys, as well as adding enter and tab to those keys.

Signed-off-by: Tametomo <irarice@gmail.com>