A few keyboard fixes, with one of them reported by Jarvellis.
Reverted back to the old check for item shortcuts, as the new method
required continuous keys(which is a reasonable assumption that the user
will do that , but is still incorrect), made emotes override item
shortcuts in priority, so that an item doesn't accidentally get used,
and moved emotes inside the game loop, since it was accidentally left
out and could cause a client crash if one of them was used outside of
the game. Also don't have the emote meta key override other key usage
when they're not used for emotes.

Signed-off-by: Tametomo <irarice@gmail.com>