Continuation of reorganization of the client folders as started earlier with the bindings folder creation. The bindings folder is to be used for containing generic wrappers to specific external libraries used within the client, as well as generic helper classes which extend or add to the functionality of that library. The core folder contains non server specific code that is independent of any specific library, and as such, contains the basic core functionalities for the client. The eathena folder contains server specific data structures, library extensions, and core classes which may or may not need to be reimplemented for different servers.
Currently, there still needs to be a lot of work done to have these
classes encapsulate these principles the best that they can, but
overall, this should help a lot more with defining the encapsulation
lines a lot better.

Also merged nearly all of the content of the ItemShortcut and
EmoteShortcut classes as well, so that the generic component could be
placed in the bindings folder, as well as reducing the amount of
redundant code and making new shortcut handler classes easier to write.

Signed-off-by: Tametomo <irarice@gmail.com>