Don't show item popups when the widget isn't visible (On rare cases, ItemContainers could gain focus when the parent window hid).
While this does fix this specific issue, what we really need is for
recursive widgetShown and widgetHidden events, in order to ensure that
all cases where a subwindow that this widget brings up are hidden on
the parent window hiding. Otherwise, tests for these cases are rather
ugly, and a bit unreliable when new use cases come up.

Signed-off-by: Tametomo <irarice@gmail.com>