Some cleanups to the trade window.
Add button is now not enabled unless OK hasn't been pressed yet, or if
there is a selected item in the inventory window (which won't select
an item if the inventory window isn't selected). Also made the money
field into an IntTextField, so that string values couldn't be sent as
money.

Signed-off-by: Tametomo <irarice@gmail.com>