Fixed up some button states so that they disable/enable as appropriate (since they don't constantly poll for changes, but instead only update on changes, some of them needed to be told to update in advance of a server acknowledgement for their action, since it currently isn't the job of the network stack to notify them that the server changed the status on the item that they previously acted on).
Also moved the add button on the Trade window to the Inventory window,
since we are indeed adding an item to trade, it's in an inconvenient
spot, since the item we want to act on is not in the Trade window, just
like the store button in the Storage window used to be cumbersome as
well. Also fixed up some trading states as well so that trading states
can be a bit more accurate.

Signed-off-by: Tametomo <irarice@gmail.com>