Modified scroll wheel event handling behavior.
Under the GUIChan scrollwheel handling, the scrollwheel will only work
for a widget if it's contained within the widget. This patch changes
that behavior so that the scroll wheel first will defer to the widget
below the current mouse pointer, like GUIChan currently does, but then
if the event is still not consumed, it is then deferred to the currently
focused widget.

What this basically means for the client is that if the mouse is not
already over a widget which can use the scrollwheel, then the focused
widget will be given a chance to use it (which can be particularly
useful if the widget it has to dwell over is particularly small, or if
the user is intending to use the scrollwheel like a key shortcut).

Signed-off-by: Tametomo <irarice@gmail.com>