Fixed mouse cursor not updating on widget shown/hidden events.
TODO: Fix the last corner case with the mouse state, which is if the two
windows overlap and the mouse is in an area which can get resized for
both windows, and the bottom widget is then hidden. This might not be
fixable until GUIChan 0.9.0, which has a method which returns all of the
widgets at the specific coordinates given.

Signed-off-by: Tametomo <irarice@gmail.com>