Download news.txt like any other update
The update downloading code had an extra code path that was only used to
download news.txt without caching it on disk.

This patch removes that code path.  Technically it wastes disk space for
a small text file, but in a cache with about 10MB of image files that's
not significant.