Fix order of calling loadUpdates()
The name of loadUpdates is causing confusion - it doesn't download
updates, it finds already-downloaded updates and adds them to
ResourceManager.

It needs to be called after the updates have downloaded, this fixes a
crash in the previous refactor preparation.  Possibly still not the
right place to call it from, but that can be improved as part of an
upcoming refactor.