Made the Reset Windows button available outside of the game as well, since there's no real reason why resetting the windows can't be done before getting in game. This removes an unneeded state reference.
Also fixed configuration saving, which was broken when the Engine was split
out, and occurred because of how gcc handles stack frames (the current state
system is a polling one, so when the state is switched, first the Exit state
would be called, then it would trace back through finishing the Game state
once that was done. Seems like a buggy behavior to me, just like how when testing
enumerations, the compiler will only test the enumeration, even if there's
succeeding arguements. The upcoming state system will not have this problem,
since it is event based, instead of polled).

Signed-off-by: Tametomo <irarice@gmail.com>