Only have one running network instance, since the client can only manage one to begin with.
If multiple connections are needed in the future (not a problem with eAthena),
then we should consider instead making the Network class manage running
connections on a port by port basis to the same address. It's better anyways
to have just one interface that manages all inbound and outbound connections,
rather than several, and running the risk of the threads running out of
sync of each other.

Signed-off-by: Tametomo <irarice@gmail.com>