Add a way to manually override whether to autofocus on the most common action for a window on opening it in the configuration file.
I personally don't consider this to be too valuable of a configuration option,
but at the same time, can see why some people would prefer to manually keep
control of their own focus. In any case, it's a configuration option that
doesn't hurt, but won't be directly exposed in the GUI, due to my opinion
of it not being particularly useful enough to justify that.

Signed-off-by: Tametomo <irarice@gmail.com>