Don't show item popups with ItemShortcutContainers.
The shortcut bars are used like a quick access bar currently, so a user is
less inclined to want to know the information about the item in those slots
without going to the Inventory, Equipment, and other item contaner screens.

Might reimplement info showing in the future in a different way by adding a
PopupMenu option to show the item info, but item popups should not show in the
ItemShortcutContainers without direct user interaction, since that appears to
be a rather undesireable behavior to some players.

Signed-off-by: Tametomo <irarice@gmail.com>