Allow for the emote shortcut bar to be filled by default again.
This was broken a while ago to allow for item and emote shortcuts
to share the exact same handler. It took so long to come up with
an acceptable way to fill the bar again because all of the solutions
being proposed at the time would have broken the simplicity of the
new combined class.

Signed-off-by: Tametomo <irarice@gmail.com>