Some changes/revamps to the DownloadWrapper and update downloader behavior.
Fixed curl leaking, made canceling not cause a client exit if it doesn't have
to, swapped loading order for news.txt and resources2.txt, in order to improve
download progress reporting (since resources2.txt is a more critical file, and
news.txt can be seen as an optional file. Loading resources2.txt first allows
for the client to know exactly how much to download before committing to
reporting it), as well as fixed download progress reporting.

Signed-off-by: Tametomo <irarice@gmail.com>