Fixed a TabbedArea regression that occurred with the merger of the state management branch. Likely was an error beforehand as well, but was likely a silent problem.
The issue fixed is that tab deletion is delayed within tabbed areas
until it is safe to do so within the client execution. Otherwise,
we'll end up with a GUIChan exception if we happen to be on a tab
that is going to be deleted, or worse, segfaulting if the tab is
currently being displayed. Delayed deletion solves both of these
potential issues.

Signed-off-by: Tametomo <irarice@gmail.com>