While this doesn't really make a difference with SDL 1.2, since it only supports one video target at the moment, make sure to store the SDL video mode to a temporary variable, then set the target after the resolution change has been confirmed to be successful.
The reason I say this doesn't really matter is that if it wasn't
recycling the same video pointer, we'd have a leak on the setTarget
function in the base Graphics class, which doesn't occur. However,
this will likely change for SDL 1.3, so it'd be better to be prepared
than not notice the problem later.

Signed-off-by: Tametomo <irarice@gmail.com>