Partial reversion/reworking of a previous commit from today.
Commit 3b2984d664cb006842afa15fc907feb6fcb247e0 added a downloadFailed()
line after telling the user that the update process is incomplete. However,
this prevents the user from being able to continue. This shouldn't occur
at this point, but rather later, when it determines that adding updates to
the resman failed (for instance, if a file is missing). This way, there
still is a chance for the updates to fail after this point, but it no longer
blocks unless it determines it's missing a file.

Signed-off-by: Tametomo <irarice@gmail.com>