Some changes to downloading behavior.
First up, curl errors are no longer reported to the command line, but
to the runtime.log file instead. This way, curl errors won't go undiscovered
with developers if an ordinary user submits their log for a bug report. If
developers still wish to see when curl errors occur, they can manually add
in a debugging statement, just like with anything else.

Next, update downloading logic was cleaned up in a few ways. First off, the
updater won't show that it's attempting to download files after a cancel any
more, but will instead attempt to load files from the state where the user
leaves it. Then, it will attempt to checksum all files listed in
resources2.txt. If the files fail to checksum, the client will then prompt
the user whether they wish to proceed, or attempt to redownload. This gets
rid of an old TODO left around by Octalot.

Also, not having a valid resources2.txt file is now a fatal error, even if the
user cancels ahead of time.

Signed-off-by: Tametomo <irarice@gmail.com>