Prevent the gcc compiler from entering into predictive branching in an area where it could be potentially hazardous to do so. While this specific area probably isn't as important to avoid the while{switch} construct as other areas are, it still isn't so nice to be running ahead of where we want the code to be running, especially for an external thread.
TODO: The friend class MessageOut could be causing some potential
threading errors when the network is switching ports on the order of
microseconds (due to a potential of overflowing mOutBuffer and
clobbering mInSize). This, as well as many other constructs within the
networking stack need to be addressed. In fact, it probably would be
best as well to be running the outgoing messages in their own thread
as well, instead of just the incoming messages, and to try to address
some of the inefficiencies/shortcomings of the current network design.

Signed-off-by: Tametomo <irarice@gmail.com>