Added death effects to particle engine.
Every particle can now have a death effect. This is an effect which
is created when the particle dies. Which death reasons (timeout,
touching floor, touching sky, reaching target or deleted by external
call) trigger the effect can also be specified. This is useful for
exploding projectiles and many other effects.

Reviewed-by: Bertram
Signed-off-by: Tametomo <irarice@gmail.com>