Some miscellaneous additions/improvements to the tabbed area.
First up, tabs which are created by a name string through addTab are
now returned on creation, as well as allowing developers to dictate
whether the tab to be created will be closeable or not. This makes
this function equal in what can be done with it to the other addTab
version, where you would supply the tab in advance.

Second, a new option has been added to the TabbedArea to allow it to
hide its tabs when there's only one tab available, and is enabled by
default. This option makes it so that interface space isn't wasted
when there's no other tabs to select.

Signed-off-by: Tametomo <irarice@gmail.com>