Some minor changes when it comes to color sanitization.
First off, when sanitizing colors, it should be completely invisible
to the user how we're sanitizing them. This patch addresses this by
ensuring that tags which don't evaluate to color tags look the same
when sanitizing/not sanitizing.

Second, we should never display text in the chat window or over the
player that evaluates to an empty string when not sanitizing the text.
As such, any string that ends up evaluating to nothing will be ignored.

Signed-off-by: Tametomo <irarice@gmail.com>