Don't use the destroy type on deleting an internal helper class within a class. This could sometimes result in causing a segmentation fault otherwise, when the internal helper class needs to call its own destructor.
TODO: The whole helper class paradigm that we're forced to use with
GUIChan currently results in a lot of similar yet different code
blocks. This seems like a lot of bureaucratic waste to me, and as
such, we should try to brainstorm some ways in which we can get
the same flexibility without relying on so much copy/paste waste.

Signed-off-by: Tametomo <irarice@gmail.com>