Partial cherrypick of ManaSource commit 91e2779891973cdbefc6e07ca3dee7a432a05112
Relevant comments from the original commit:

Implemented a screen shake effect system in the viewport class.

The screen can either be "nudged" in a random direction with a specific
intensity or you can define an exact x and y intensity, decay factor and
duration. A method for stopping all shake effects is also implemented, but
not used yet.

Differences for this commit:

Linked the original changes through the effect system, so that external
effects can choose to use quake effects.

TODO: Allow for ambient effects to be applied through this manner as well.

Signed-off-by: Tametomo <irarice@gmail.com>