Minor commit. Change all SDL integer types into C99 standard integer types instead for future proofing the client, in case SDL is not used later on.
We use the stdint.h C99 header instead of c++0x's cstdint, so that
the maximal compiler support is available. Since c++0x is not an
official standard quite yet, and that not using it makes it so that
more compiler choices are available for use, we're avoiding c++0x
until such time in which it is widely adopted, and we actually have
a need for it. ;)

TODO: Manually review each instance of an non-architecture standard
      integer and ensure that it actually needs to be that size.

Signed-off-by: Tametomo <irarice@gmail.com>