Added window resizing support.
This was previously ignored before, because we didn't know what sort of
drawing behavior an SDL_RESIZE event might cause (and didn't want to
implement it if it would cause poor CPU performance), and didn't want to
invest a lot of time into investigating it for this release.

The concerns however proved to be unfounded though, and minimal changes were
needed to implement this. As such, this poses no more risk to adding testing
complexity more than the window resizing code already does, since it's hooked
in the same methods. But to be sure that risks were kept minimal, I am
currently requiring that if you are to switch to fullscreen mode, you need to
use one of the existing modes, in case custom modes causes a problem on
Windows in a nonvirtualized environment.

On a side note though, the client can be maximized in Windows, but X11
technically has no such ability to do so. Further investigation into how to use
EWMH hints like the _NET_WM_STATE_MAXIMIZED_VERT, or _NET_WM_STATE_MAXIMIZED_HORZ
atoms might bring maximize functionality to X11 as well, but are probably
more time than they are worth right now.

As for OS X, there's no sense in implementing a maximize functionality there,
when a zoom interface would make more sense. But without having a test
platform for this, or understanding how to best interact with Carbon to do so,
then this will also have to wait.

Signed-off-by: Tametomo <irarice@gmail.com>